+++ packages/server/src/@trpc-server/http.ts
@@ -1,0 +1,2 @@
+// This should be moved to a like `@trpc/plugin`-package
+export * from '@trpc/core/http';
