+++ packages/server/src/adapters/aws-lambda/utils.ts
@@ -1,5 +1,3 @@
-import type { AnyRouter, inferRouterContext } from '@trpc/core';
-import { TRPCError } from '@trpc/core';
 import type {
   APIGatewayProxyEvent,
   APIGatewayProxyEventV2,
@@ -7,6 +5,8 @@ import type {
   APIGatewayProxyStructuredResultV2,
   Context as APIGWContext,
 } from 'aws-lambda';
+import type { AnyRouter, inferRouterContext } from '../../@trpc-server'; // import @trpc/server
+import { TRPCError } from '../../@trpc-server'; // @trpc/server
 import type {
   HTTPHeaders,
   OnErrorFunction,
