+++ packages/server/src/adapters/fastify/fastifyRequestHandler.ts
@@ -1,6 +1,6 @@
 import { Readable } from 'node:stream';
-import type { AnyRouter } from '@trpc/core';
 import type { FastifyReply, FastifyRequest } from 'fastify';
+import type { AnyRouter } from '../../@trpc-server'; // @trpc/server
 import type {
   HTTPBaseHandlerOptions,
   HTTPRequest,
