+++ packages/server/src/adapters/fetch/fetchRequestHandler.ts
@@ -1,4 +1,4 @@
-import type { AnyRouter } from '@trpc/core';
+import type { AnyRouter } from '../../@trpc-server'; // @trpc/server
 import type {
   HTTPRequest,
   HTTPResponse,
