+++ packages/server/src/adapters/node-http/nodeHTTPRequestHandler.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
-import type { AnyRouter } from '@trpc/core';
+import type { AnyRouter } from '../../@trpc-server'; // @trpc/server
+
 import type {
   HTTPRequest,
   HTTPResponse,
