+++ packages/server/src/adapters/node-http/types.ts
@@ -1,6 +1,12 @@
 import type { IncomingMessage, ServerResponse } from 'http';
-import type { AnyRouter, inferRouterContext, MaybePromise } from '@trpc/core';
-import type { HTTPBaseHandlerOptions, TRPCRequestInfo } from '../../http';
+// eslint-disable-next-line no-restricted-imports
+import type { MaybePromise } from '@trpc/core';
+import type { AnyRouter, inferRouterContext } from '../../@trpc-server'; // @trpc/server
+import type {
+  HTTPBaseHandlerOptions,
+  TRPCRequestInfo,
+} from '../../@trpc-server/http';
+// @trpc/server/http
 import type { NodeHTTPContentTypeHandler } from './internals/contentType';
 
 interface ParsedQs {
