+++ packages/server/src/http.ts
@@ -1,2 +1,1 @@
-// This should be moved to a like `@trpc/plugin`-package
-export * from '@trpc/core/http';
+export * from './@trpc-server/http';
