+++ packages/server/src/index.ts
@@ -1,35 +1,1 @@
-export {
-  TRPCError,
-  /**
-   * @deprecated use `experimental_trpcMiddleware` instead
-   */
-  experimental_standaloneMiddleware,
-  experimental_standaloneMiddleware as experimental_trpcMiddleware,
-  initTRPC,
-  type inferProcedureInput,
-  type inferProcedureOutput,
-  type inferRouterError,
-  type inferRouterInputs,
-  type inferRouterOutputs,
-  type AnyProcedure as AnyTRPCProcedure,
-  type AnyRouter as AnyTRPCRouter,
-} from './unstableDoNotImportThis';
-
-export type {
-  /**
-   * @deprecated use `AnyTRPCProcedure` instead
-   */
-  AnyProcedure,
-  /**
-   * @deprecated use `AnyTRPCRouter` instead
-   */
-  AnyRouter,
-  AnyMiddlewareFunction,
-} from '@trpc/core';
-
-/**
- * @deprecated
- * Use `Awaited<ReturnType<typeof myFunction>>` instead
- */
-export type inferAsyncReturnType<TFunction extends (...args: any[]) => any> =
-  Awaited<ReturnType<TFunction>>;
+export * from './@trpc-server';
